+++ CHANGES.markdown
@@ -26,9 +26,10 @@ None
 * Windows 5.03 Hop Boil Times [1082](https://github.com/Brewtarget/brewtarget/issues/1082)
 * Deleting recipe locks associated linked processes from deletion [1085](https://github.com/Brewtarget/brewtarget/issues/1085)
 * Stock Purchase button showing on incorrect catalog pages [1086](https://github.com/Brewtarget/brewtarget/issues/1086)
+* --create-blank cli option exits with segfault [1088](https://github.com/Brewtarget/brewtarget/issues/1088)
 
 ### Release Timestamp
-Sun, 15 Mar 2026 05:00:04 +0100
+Tue, 17 Mar 2026 05:00:04 +0100
 
 ## v5.0.3
 Bug fixes and minor enhancements.